fix: improve toast notifications, loading states, and test coverage
- Fix toast notification types: use 'error' instead of 'warning' for verification errors
- Consolidate toast messages to single string without description object
- Extract verify-email page client logic to separate VerifyEmailClient component
- Add Suspense boundary for verify-email page
- Fix SystemStatusBanner loading state: move statusConfig memo before conditional check
- Add proper loading state for uptime stats with dedicated isFetchingUptime flag
- Add uptime error state in health slice for better error handling
- Update useUptimeStats hook to use dedicated uptime loading and error states
- Fix polling interval from 30s to 60s in tests to match HEALTH_CONSTANTS
- Add fetchUptimeStats mock in SystemStatusBanner tests

Files affected:
- LoginClient.tsx: fix toast types for verification errors
- RegisterClient.tsx: consolidate toast message format
- verify-email/page.tsx: extract client logic, add Suspense
- VerifyEmailClient.tsx: new file with verification logic
- SystemStatusBanner.tsx: move statusConfig memo before conditional
- useUptimeStats.ts: use dedicated uptime loading/error states
- health.types.ts: add isFetchingUptime and uptime error fields
- healthSlice.ts: add dedicated uptime loading and error state management
- SystemStatusBanner.test.tsx: update polling interval and add mocks
- package-lock.json: update dependencies
